Create di.brilliance.r
Dichromatism Measurement3 : Brilliance Comparison

Measures the difference in overall brilliance between male and female birds of a species. When log-avg option is FALSE (default), the measurement compares the average of the sum of relative reflectance over the entire spectral range (the integral under the curve) for all reflectance curves in the male or female individuals. When log-avg option is TRUE (must set), the function follows Maia et al. 2013 to compare the log-scaled average reflectance value across all spectra.